+++ lib/src/_modifiers.dart
@@ -3,9 +3,9 @@ import 'package:domino/domino.dart';
 // https://bulma.io/documentation/modifiers/syntax/
 
 class Mods {
-  final MainColor _mainColor;
-  final SizeMod _sizeMod;
-  final FloatMod _floatMod;
+  final String _color;
+  final String _size;
+  final String _float;
   final bool _outlined;
   final bool _loading;
   final bool _disabled;
@@ -14,9 +14,9 @@ class Mods {
   final bool _clipped;
 
   Mods({
-    MainColor mainColor,
-    SizeMod sizeMod,
-    FloatMod floatMod,
+    String color,
+    String size,
+    String float,
     bool outlined,
     bool loading,
     bool disabled,
@@ -24,9 +24,9 @@ class Mods {
     bool paddingless,
     bool clipped,
   })
-      : _mainColor = mainColor,
-        _sizeMod = sizeMod,
-        _floatMod = floatMod,
+      : _color = color,
+        _size = size,
+        _float = float,
         _outlined = outlined,
         _loading = loading,
         _disabled = disabled,
@@ -36,9 +36,9 @@ class Mods {
 
   Element apply(Element element) {
     element
-      ..addClass(mainColorClass(_mainColor))
-      ..addClass(sizeModClass(_sizeMod))
-      ..addClass(floatModClass(_floatMod))
+      ..addClass(_color)
+      ..addClass(_size)
+      ..addClass(_float)
       ..addClass(outlinedClass(_outlined))
       ..addClass(loadingClass(_loading))
       ..addClass(marginlessClass(_marginless))
@@ -49,53 +49,25 @@ class Mods {
   }
 }
 
-enum MainColor { primary, link, info, success, warning, danger }
-enum SizeMod { small, medium, large }
-enum FloatMod { clearfix, pulledLeft, pulledRight }
-
-String mainColorClass(MainColor mainColor) {
-  if (mainColor == null) return null;
-  switch (mainColor) {
-    case MainColor.primary:
-      return 'is-primary';
-    case MainColor.link:
-      return 'is-link';
-    case MainColor.info:
-      return 'is-info';
-    case MainColor.success:
-      return 'is-success';
-    case MainColor.warning:
-      return 'is-warning';
-    case MainColor.danger:
-      return 'is-danger';
-  }
-  return null;
+abstract class Color {
+  static const String primary = 'is-primary';
+  static const String link = 'is-link';
+  static const String info = 'is-info';
+  static const String success = 'is-success';
+  static const String warning = 'is-warning';
+  static const String danger = 'is-danger';
 }
 
-String sizeModClass(SizeMod sizeMod) {
-  if (sizeMod == null) return null;
-  switch (sizeMod) {
-    case SizeMod.small:
-      return 'is-small';
-    case SizeMod.medium:
-      return 'is-medium';
-    case SizeMod.large:
-      return 'is-large';
-  }
-  return null;
+abstract class Size {
+  static const String small = 'is-small';
+  static const String medium = 'is-medium';
+  static const String large = 'is-large';
 }
 
-String floatModClass(FloatMod floatMod) {
-  if (floatMod == null) return null;
-  switch (floatMod) {
-    case FloatMod.clearfix:
-      return 'is-clearfix';
-    case FloatMod.pulledLeft:
-      return 'is-pulled-left';
-    case FloatMod.pulledRight:
-      return 'is-pulled-right';
-  }
-  return null;
+abstract class Float {
+  static const String clearfix = 'is-clearfix';
+  static const String pulledLeft = 'is-pulled-left';
+  static const String pulledRight = 'is-pulled-right';
 }
 
 String _class(bool boolValue, String className) =>